+++ test/amount_rate_test.dart
@@ -1,0 +1,21 @@
+import 'package:test/test.dart';
+import 'package:forestvpn_api/forestvpn_api.dart';
+
+// tests for AmountRate
+void main() {
+  final instance = AmountRateBuilder();
+  // TODO add properties to the builder and call build()
+
+  group(AmountRate, () {
+    // String currency
+    test('to test the property `currency`', () async {
+      // TODO
+    });
+
+    // double amount
+    test('to test the property `amount`', () async {
+      // TODO
+    });
+
+  });
+}